requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.5
pillow==10.4.0
pytesseract==0.3.10
faker==27.0.0
pydantic-settings==2.4.0
httpx==0.27.0
pymupdf==1.24.9
img2pdf==0.5.1
opencv-python-headless>=4.9.0
numpy>=1.26